[Test][Isolated Region] Reduce memory usage in test_slurm_memory_based_scheduling
The changed line is running on c5.large, which has 4GB of physical memory, when running with 3GB of program, it failed in isolated region on RHEL8. Slightly reduce the memory usage to 2.8GB to make the test pass. ParallelCluster test shouldn't check too strictly about memory usage because it is managed by operating systems.