[integ-tests] In EFA test, use the same instance type for head node and compute nodes when possible
This commit uses the same instance type for both compute node and head node when the instance type is available in open capacity pool. When the instance type is p4, p5, p6, or hpc which require capacity reservations or capacity blocks, the head node still uses a different instance type as the logic before this commit.

Prior to this commit, there were intermittent failures of insufficient capacity of the head node instance type.

Although the head node is not in the placement group, which is used by the integration tests framework to make the capacity reservation, this commit "hopes" that the AZ has adequate capacity of the instance type. In other words, if the integration test framework can make a capacity reservation of 35 instances in a placement group in the AZ, this commit hopes there will be an additional 1 instance anywhere in the AZ.